Moved from separated tab to a part of client profile list
