+++ Mobile/CarWashService/Views/MainView.swift
@@ -9,7 +9,7 @@
 import SwiftUI
 
 struct MainView: View {
-    @State private var selection = 2
+    @State private var selection = 1
 
     var body: some View {
         TabView(selection: $selection) {
@@ -18,16 +18,11 @@ struct MainView: View {
                         Image(systemName: "person.fill")
                         Text("Profile")
                     }.tag(0)
-            ClientCarsView()
-                    .tabItem {
-                        Image(systemName: "car.fill")
-                        Text("Cars")
-                    }.tag(1)
             CarWashesView()
                     .tabItem {
                         Image(systemName: "wand.and.rays")
                         Text("Car Washes")
-                    }.tag(2)
+                    }.tag(1)
         }.accentColor(ApplicationColor.Primary.toRGB())
     }
 }
